Added tool based on libtooling presentation.
This is just a dummy tool at the moment using the latest
version of Clang to replace appropriate declarations
and calls of "Get" to "front" as per the libtooling
presentation at BoostCon.  Pretty much a version of
Chandler Carruth's example, but brought up to date to
work with the latest Clang.